+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.14.38
+
+* added `call_between_retries` to `retry`